DT-400 Move the allocate POM integration tests to their rightful home (#107)
